Fixed broken test createProvider
Moved createProvider to its own module
